website_status_checker/requirements.txt
python-dotenv==1.1.0
requests==2.32.3

